Moving functions to registry.h and renaming to following existing naming patterns.
