+++ tests/cpplatest/CMakeLists.txt
@@ -18,7 +18,7 @@ endif()
 target_sources(witest.cpplatest PUBLIC
     ${COMMON_SOURCES}
     ${COROUTINE_SOURCES}
-    ${CMAKE_CURRENT_SOURCE_DIR}/../BasicRegistryTests.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/../RegistryTests.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/../CppWinRTTests.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/../CppWinRT20Tests.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/../StlTests.cpp
